DesafioApp/requirements.txt
asgiref==3.8.1
Django==5.0.7
sqlparse==0.5.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
markdown==3.6
django-filter==24.2
djangorestframework-simplejwt==5.3.1
django-compressor==4.5.1